Move caption keyboard shortcut tests into attachments test suite
Consolidate the standalone bug reproduction tests into the existing
attachments.test.js for a cohesive, well-organized test suite.
